[web] Detach `MDXRenderer` from `BlogPage`
