+++ apps/web/components/mdx-renderer/index.ts
@@ -1,0 +1,1 @@
+export { MDXRenderer } from './MDXRenderer';
